Bhagyashree Mangnale
Being a fresher, I think I am very flexible and adaptive to learning new things. I am sure I will be able to contribute something capable for the growth of the company. My last project in Operations has taught me how to be a team player, and work in unison. I am sure that is a valuable quality which I will be bringing to the table. I will follow the company guidelines and always play by the rulebooks.